BOTH: Various small fixes
- fixed many wrong token checks
- fixed "don't open menu during load" fix (originally one check omitted)
- removed -ffast-math compile option from Release
- added -ggdb to copmile options of Debug (otherwise it is missing symbols for debugging with GDB)
- added code for striping all symbols from Release targets

+++ SP/code/cgame/cg_effects.c
@@ -983,7 +983,7 @@ void CG_SparklerSparks( vec3_t origin, int count ) {
 
 // just a bunch of numbers we can use for pseudo-randomizing based on time
 #define NUMRANDTABLE 257
-unsigned int randtable[NUMRANDTABLE] =
+unsigned short randtable[NUMRANDTABLE] =
 {
 	0x0000, 0x1021, 0x2042, 0x3063, 0x4084, 0x50a5, 0x60c6, 0x70e7,
 	0x8108, 0x9129, 0xa14a, 0xb16b, 0xc18c, 0xd1ad, 0xe1ce, 0xf1ef,
